Deleted redundant layout folder and moved all my new code to features/common/cy
