--- src/client/features/view/components/graph/layout/coseBilkent.js
@@ -1,13 +1,0 @@
-const coseBilkentOpts = {
-  name: 'cose-bilkent',
-  fit: true,
-  randomize: false,
-  animate: 'end',
-  animationEasing: 'ease-out-cubic',
-  animationDuration: 800,
-  tilingPaddingVertical: 20,
-  tilingPaddingHorizontal: 20,
-  nodeDimensionsIncludeLabels: true
-};
-
-module.exports = coseBilkentOpts;
