--- src/client/features/view/components/graph/layout/dagre.js
@@ -1,6 +1,0 @@
-const dagreOpts = {
-  name: 'dagre',
-  rankDir: 'LR'
-};
-
-module.exports = dagreOpts;
