--- src/client/features/view/components/graph/layout/klay.js
@@ -1,16 +1,0 @@
-const klayOpts = {
-  name: 'klay',
-  klay: {
-    borderSpacing: 10,
-    separateConnectedComponents: true,
-    thoroughness: 100,
-    compactComponents: false,
-    spacing: 40,
-    edgeSpacingFactor: 2.0,
-
-    layoutHierarchy: true
-  },
-  nodeDimensionsIncludeLabels: false
-};
-
-module.exports = klayOpts;
