--- src/client/features/view/components/graph/layout/preset.js
@@ -1,12 +1,0 @@
-const presetOpts = (layoutJSON) => {
-  return {
-    name: 'preset',
-    positions: node => {
-      return layoutJSON[node];
-    }, 
-    fit: true,
-    padding: 50
-  };
-};
-
-module.exports = presetOpts;
